bad words

+++ LibraryModule/CommentaireService.js
@@ -1,38 +1,66 @@
 var express = require('express');
 var Commentaire = require ("./CommentaireModel")
 var Library  =require("./LibraryModel")
+const validatorRegisteeer = require("./validation/ValidCmntr");
+const badWords = require('bad-words');
+const https = require('https');
+const csv = require('csv-parser');
 
-async function addC(req,res,next){
- 
+function filterWords(text, badWords) {
+  if (!badWords || badWords.length === 0) {
+    return text;
+  }
 
+  for (let i = 0; i < badWords.length; i++) {
+    const regex = new RegExp(`\\b${badWords[i]}\\b`, 'gi');
+    text = text.replace(regex, '*'.repeat(badWords[i].length));
+  }
+
+  return text;
+}
+
+async function addC(req, res, next) {
+  const { errors, isValid } = validatorRegisteeer(req.body);
+
+  console.log(req.body)
+  console.log(isValid)
+
+  if (isValid == true) {
     let newCommentaire;
-    try{
-      const data = Commentaire(
-     {
 
-        description:req.body.description,
-        dateEnvoi :Date.now(),
-        Libraryid:req.body.Libraryid
+    try {
+      const badWords = ['bad', 'words', 'here','merde'];
+      const filteredText = filterWords(req.body.description, badWords);
 
-      
-    });    
-    newCommentaire = await data.save();
-    const library = await Library.findById(req.body.Libraryid);
-    if (library) {
-      library.commentaires.push(newCommentaire._id);
-      const newLibrary = await library.save();
-      res.json(data);
-    } else {
-      res.status(404).send("Library not found");
+      const data = Commentaire({
+        description: filteredText,
+        dateEnvoi: Date.now(),
+        Libraryid: req.body.Libraryid,
+      });
+
+      newCommentaire = await data.save();
+
+      const library = await Library.findById(req.body.Libraryid);
+      if (library) {
+        library.commentaires.push(newCommentaire._id);
+        const newLibrary = await library.save();
+        res.json(data);
+      } else {
+        res.status(404).send("Library not found");
+      }
+    } catch (e) {
+      res.status(500).json(e);
+      console.log(e);
+      if (newCommentaire) {
+        newCommentaire.delete();
+      }
     }
-  } catch (e) {
-    res.status(500).json(e);
-    console.log(e);
-    if (newCommentaire) {
-      newCommentaire.delete();
-    }
+  } else {
+    console.log(errors)
+    return res.status(403).json(errors);
   }
 }
+
     async function updateC(req,res,next)
  {
     Commentaire.findByIdAndUpdate(req.params.id,{
@@ -59,15 +87,23 @@ async function addC(req,res,next){
      async function deletec(req, res, next) {
     
     
-        const id = req.params.id;
-  try {
-    await Commentaire.deleteOne({ _id: id }); // delete the comment with the given ID
-    res.sendStatus(204); // send a "no content" response if the comment was successfully deleted
-  } catch (err) {
-    console.error(err);
-    res.sendStatus(500); // send a "server error" response if there was a problem deleting the comment
-  }
-}
+      try {
+        const commentaire = await Commentaire.findById(req.params.id);
+        const library = await Library.findByIdAndUpdate(commentaire.Libraryid, {
+          $pull: {
+            commentaires: {
+              _id: commentaire._id,
+            },
+          },
+        });
+        await commentaire.remove();
+        // course.lessons.splice(course.lessons.findIndex(a => a._id === lesson._id) , 1)
+        res.status(200).json("commentaire deleted");
+      } catch (error) {
+        res.json(error);
+      }
+    };
+
 
     
     module.exports={addC,updateC,listC,deletec}
